+++ xeofs/__init__.py
@@ -1,4 +1,4 @@
 from importlib.metadata import version
-from xeofs import models, xarray
+from xeofs import models
 
 __version__ = version('xeofs')
