+++ xeofs/models/mca.py
@@ -1,187 +1,312 @@
-from typing import Optional, Union, Iterable, Tuple, List
-
 import numpy as np
+import xarray as xr
 
-from xeofs.models._base_mca import _BaseMCA
-from xeofs.models._transformer import _MultiArrayTransformer
-from ..utils.tools import squeeze
-
-import warnings
-
-warnings.filterwarnings("ignore", message="numpy.dtype size changed")
-warnings.filterwarnings("ignore", message="numpy.ufunc size changed")
-
-Array = np.ndarray
-ArrayList = Union[Array, List[Array]]
+from ._base_cross_model import _BaseCrossModel
+from .decomposer import CrossDecomposer
+from ..utils.data_types import XarrayData, DataArrayList
+from ..utils.tools import _hilbert_transform_with_padding
+from ..utils.testing import pearson_correlation
 
 
-class MCA(_BaseMCA):
-    '''Maximum Covariance Analysis (MCA) of a single or multiple ``np.ndarray``.
+class MCA(_BaseCrossModel):
 
-    MCA is also known as Singular Value Decomposition (SVD) analysis or
-    Partial Least Squares (PLS) analysis.
+    def fit(self, data1: XarrayData | DataArrayList, data2: XarrayData | DataArrayList, dims, weights1=None, weights2=None):
+        '''
+        Fit the model.
 
-    Parameters
-    ----------
-    X, Y : np.ndarray
-        Data matrices two compute maximum covariance. ``X`` and ``Y`` can be
-        any N-dimensional array. Dimensions along which covariance shall be
-        maximised (denoted as *samples*) have to be defined by the
-        ``axis`` parameter. All remaining axes will be reshaped into a new
-        axis called *features*. Sample dimension of ``X`` and ``Y`` must be
-        the same.
-    n_modes : Optional[int]
-        Number of modes to compute. Computing less modes can results in
-        performance gains. If None, then the maximum number of modes is
-        equivalent to ``min(n_samples, n_features)`` (the default is None).
-    norm : bool
-        Normalize each feature (e.g. grid cell) by its temporal standard
-        deviation (the default is False).
-    weights_X : Optional[np.ndarray] = None
-        Weights applied to features of ``X``. Must have the same dimensions as the
-        original features which are the remaining axes not specified by
-        ``axis`` parameter).
-    weights_Y : Optional[np.ndarray] = None
-        Weights applied to features of ``Y``. Must have the same dimensions as the
-        original features which are the remaining axes not specified by
-        ``axis`` parameter).
-    axis : Union[int, Iterable[int]]
-        Axis along which variance should be maximised. Can also be
-        multi-dimensional. For example, given a data array of dimensions
-        ``(n x p1 x p2)`` with `n` time series at ``p1`` and ``p2`` different
-        locations, ``axis=0`` will maximise `temporal` variance along ``n``.
-        In contrast, ``axis=[1, 2]`` will maximise `spatial` variance along
-        ``(p1 x p2)`` (the default is 0).
+        Parameters:
+        -------------
+        data1: xr.DataArray or list of xarray.DataArray
+            Left input data.
+        data2: xr.DataArray or list of xarray.DataArray
+            Right input data.
+        dims: tuple
+            Tuple specifying the sample dimensions. The remaining dimensions 
+            will be treated as feature dimensions.
+        weights1: xr.DataArray or xr.Dataset or None, default=None
+            If specified, the left input data will be weighted by this array.
+        weights2: xr.DataArray or xr.Dataset or None, default=None
+            If specified, the right input data will be weighted by this array.
 
+        '''
+        self._preprocessing(data1, data2, dims, weights1, weights2)
 
-    Examples
-    --------
+        decomposer = CrossDecomposer(n_components=self._params['n_modes'])
+        decomposer.fit(self.data1, self.data2)
 
-    Import package and create data:
+        # Note:
+        # - explained variance is given by the singular values of the SVD;
+        # - We use the term singular_values_pca as used in the context of PCA:
+        # Considering data X1 = X2, MCA is the same as PCA. In this case,
+        # singular_values_pca is equivalent to the singular values obtained
+        # when performing PCA of X1 or X2.
+        self._singular_values = decomposer.singular_values_
+        self._explained_variance = decomposer.singular_values_  # actually covariance
+        self._squared_total_variance = decomposer.squared_total_variance_
+        self._squared_covariance_fraction = self._explained_variance**2 / self._squared_total_variance
+        self._singular_values_pca = np.sqrt(self._singular_values * (self.data1.shape[0] - 1))
+        self._singular_vectors1 = decomposer.singular_vectors1_
+        self._singular_vectors2 = decomposer.singular_vectors2_
+        self._norm1 = np.sqrt(self._singular_values)
+        self._norm2 = np.sqrt(self._singular_values)
 
-    >>> import numpy as np
-    >>> from xeofs.models import MCA
-    >>> rng = np.random.default_rng(7)
-    >>> X = rng.standard_normal((45, 4, 3))
-    >>> Y = rng.standard_normal((45, 2, 5))
+        self._explained_variance.name = 'explained_variance'
+        self._squared_total_variance.name = 'squared_total_variance'
+        self._squared_covariance_fraction.name = 'squared_covariance_fraction'
+        self._norm1.name = 'left_norm'
+        self._norm2.name = 'right_norm'
 
-    Initialize standardized MCA analysis and compute the first 2 modes:
+        # Project the data onto the singular vectors
+        sqrt_expvar = np.sqrt(self._explained_variance)
+        self._scores1 = xr.dot(self.data1, self._singular_vectors1) / sqrt_expvar
+        self._scores2 = xr.dot(self.data2, self._singular_vectors2) / sqrt_expvar
 
-    >>> model = MCA(X, norm=True, n_modes=2, axis=0)
-    >>> model.solve()
+    def transform(self, **kwargs):
+        '''Project new unseen data onto the singular vectors.
 
-    Get singular vectors:
+        Parameters:
+        -------------
+        data1: xr.DataArray or list of xarray.DataArray
+            Left input data. Must be provided if `data2` is not provided.
+        data2: xr.DataArray or list of xarray.DataArray
+            Right input data. Must be provided if `data1` is not provided.
 
-    >>> model.singular_vectors()
+        Returns:
+        ----------
+        scores1: DataArray | Dataset | List[DataArray]
+            Left scores.
+        scores2: DataArray | Dataset | List[DataArray]
+            Right scores.
 
-    Get homogeneous patterns:
+        '''
+        results = []
+        if 'data1' in kwargs:
+            data1 = kwargs['data1']
+            data1 = self.scaler1.transform(data1)
+            data1 = self.stacker1.transform(data1) # type: ignore
+            scores1 = xr.dot(data1, self._singular_vectors1) / self._norm1
+            scores1 = self.stacker1.inverse_transform_scores(scores1)
+            results.append(scores1)
 
-    >>> model.homogeneous_patterns()
+        if 'data2' in kwargs:
+            data2 = kwargs['data2']
+            data2 = self.scaler2.transform(data2)
+            data2 = self.stacker2.transform(data2) # type: ignore
+            scores2 = xr.dot(data2, self._singular_vectors2) / self._norm2
+            scores2 = self.stacker2.inverse_transform_scores(scores2)
+            results.append(scores2)
 
-    '''
-    def __init__(
-        self,
-        X: ArrayList,
-        Y: ArrayList,
-        axis : Union[int, Iterable[int]] = 0,
-        n_modes : Optional[int] = None,
-        norm : bool = False,
-        weights_X : Optional[ArrayList] = None,
-        weights_Y : Optional[ArrayList] = None,
-    ):
+        return results
 
-        self._tfx = _MultiArrayTransformer()
-        self._tfy = _MultiArrayTransformer()
-        X = self._tfx.fit_transform(X, axis=axis)
-        Y = self._tfy.fit_transform(Y, axis=axis)
-        weights_X = self._tfx.transform_weights(weights_X)
-        weights_Y = self._tfy.transform_weights(weights_Y)
+    def inverse_transform(self, mode):
+        '''Reconstruct the original data from transformed data.
 
-        super().__init__(
-            X=X,
-            Y=Y,
-            n_modes=n_modes,
-            norm=norm,
-            weights_X=weights_X,
-            weights_Y=weights_Y
-        )
+        Parameters:
+        -------------
+        mode: scalars, slices or array of tick labels.
+            The mode(s) used to reconstruct the data. If a scalar is given,
+            the data will be reconstructed using the given mode. If a slice
+            is given, the data will be reconstructed using the modes in the
+            given slice. If a array is given, the data will be reconstructed
+            using the modes in the given array.
 
-    def singular_values(self) -> Array:
-        return super().singular_values()
+        Returns:
+        ----------
+        Xrec1: DataArray | Dataset | List[DataArray]
+            Reconstructed data of left field.
+        Xrec2: DataArray | Dataset | List[DataArray]
+            Reconstructed data of right field.
 
-    def explained_covariance(self) -> Array:
-        return super().explained_covariance()
+        '''
+        svecs1 = self._singular_vectors1.sel(mode=mode)  # type: ignore
+        svecs2 = self._singular_vectors2.sel(mode=mode)  # type: ignore
 
-    def squared_covariance_fraction(self) -> Array:
-        return super().squared_covariance_fraction()
+        scores1 = self._scores1.sel(mode=mode)  # type: ignore
+        scores2 = self._scores2.sel(mode=mode)  # type: ignore
 
-    def singular_vectors(
-            self, scaling : int = 0
-    ) -> Tuple[ArrayList, ArrayList]:
-        Vx, Vy = super().singular_vectors(scaling=scaling)
-        Vx = self._tfx.back_transform_eofs(Vx)
-        Vy = self._tfy.back_transform_eofs(Vy)
-        return squeeze(Vx), squeeze(Vy)
+        Xrec1 = svecs1.dot(scores1.conj().T)
+        Xrec2 = svecs2.dot(scores2.conj().T)
 
-    def pcs(self, scaling : int = 0) -> Tuple[ArrayList, ArrayList]:
-        Ux, Uy = super().pcs(scaling=scaling)
-        Ux = self._tfx.back_transform_pcs(Ux)
-        Uy = self._tfy.back_transform_pcs(Uy)
-        return Ux, Uy
+        Xrec1 = self.stacker1.inverse_transform_data(Xrec1)
+        Xrec2 = self.stacker2.inverse_transform_data(Xrec2)
 
-    def homogeneous_patterns(self) -> Tuple[ArrayList, ArrayList]:
-        hom_pats, pvals = super().homogeneous_patterns()
-        hom_patsx = squeeze(self._tfx.back_transform_eofs(hom_pats[0]))
-        hom_patsy = squeeze(self._tfy.back_transform_eofs(hom_pats[1]))
-        pvalsx = squeeze(self._tfx.back_transform_eofs(pvals[0]))
-        pvalsy = squeeze(self._tfy.back_transform_eofs(pvals[1]))
-        return (hom_patsx, hom_patsy), (pvalsx, pvalsy)
+        Xrec1 = self.scaler1.inverse_transform(Xrec1)  # type: ignore
+        Xrec2 = self.scaler2.inverse_transform(Xrec2)  # type: ignore
 
-    def heterogeneous_patterns(self) -> Tuple[ArrayList, ArrayList]:
-        het_pats, pvals = super().heterogeneous_patterns()
-        het_patsx = squeeze(self._tfx.back_transform_eofs(het_pats[0]))
-        het_patsy = squeeze(self._tfy.back_transform_eofs(het_pats[1]))
-        pvalsx = squeeze(self._tfx.back_transform_eofs(pvals[0]))
-        pvalsy = squeeze(self._tfy.back_transform_eofs(pvals[1]))
-        return (het_patsx, het_patsy), (pvalsx, pvalsy)
+        return Xrec1, Xrec2
 
-    def reconstruct_XY(
-        self,
-        mode : Optional[Union[int, List[int], slice]] = None
-    ) -> Tuple[ArrayList, ArrayList]:
-        Xrec, Yrec = super().reconstruct_XY(mode)
-        Xrec = self._tfx.back_transform(Xrec)
-        Yrec = self._tfy.back_transform(Yrec)
-        return squeeze(Xrec), squeeze(Yrec)
+    def singular_values(self):
+        '''Return the singular values of the model.
 
-    def project_onto_left_singular_vectors(
-        self,
-        X : ArrayList = None,
-        scaling : int = 0
-    ) -> Array:
-        # Transform data to 2D
-        projx = _MultiArrayTransformer()
-        X_proj = projx.fit_transform(X, axis=self._tfx.axis_samples)
-        # Perform projection
-        pcs_X = super().project_onto_left_singular_vectors(
-            X=X_proj, scaling=scaling
-        )
-        # Backtransform to PC format
-        pcs_X = projx.back_transform_pcs(pcs_X)
-        return pcs_X
+        Returns:
+        ----------
+        singular_values: DataArray
+            Singular values of the fitted model.
 
-    def project_onto_right_singular_vectors(
-        self,
-        Y : ArrayList = None,
-        scaling : int = 0
-    ) -> Array:
-        # Transform data to 2D
-        projy = _MultiArrayTransformer()
-        Y_proj = projy.fit_transform(Y, axis=self._tfy.axis_samples)
-        # Perform projection
-        pcs_Y = super().project_onto_right_singular_vectors(
-            Y=Y_proj, scaling=scaling
-        )
-        # Backtransform to PC format
-        pcs_Y = projy.back_transform_pcs(pcs_Y)
-        return pcs_Y
+        '''
+        return self._singular_values
+    
+    def explained_variance(self):
+        '''Return explained variance.'''
+        return self._explained_variance
+    
+    def squared_covariance_fraction(self):
+        '''Return the squared covariance fraction.
+        
+        The squared covariance fraction (SCF) is the ratio of the squared covariance
+        to the total squared variance and is a measure of importance of the
+        mode. It is the analogue of the explained variance ratio in PCA.
+
+        The SCF for mode `i` is given by:
+
+        .. math::
+            SCF_i = \\frac{\\sigma_i^2}{\\sum_{i=1}^{n} \\sigma_i^2}
+
+        where :math:`\\sigma_i` is the singular value of mode `i`.
+
+        '''
+        return self._squared_covariance_fraction
+    
+    def components(self):
+        '''Return the singular vectors of the left and right field.
+        
+        Returns:
+        ----------
+        components1: DataArray | Dataset | List[DataArray]
+            Left components of the fitted model.
+        components2: DataArray | Dataset | List[DataArray]
+            Right components of the fitted model.
+
+        '''
+        svecs1 = self.stacker1.inverse_transform_components(self._singular_vectors1)
+        svecs2 = self.stacker2.inverse_transform_components(self._singular_vectors2)
+        return svecs1, svecs2
+    
+    def scores(self):
+        '''Return the scores of the left and right field.
+
+        The scores in MCA are the projection of the data matrix onto the
+        singular vectors of the cross-covariance matrix.
+        
+        Returns:
+        ----------
+        scores1: DataArray | Dataset | List[DataArray]
+            Left scores.
+        scores2: DataArray | Dataset | List[DataArray]
+            Right scores.
+
+        '''
+        scores1 = self.stacker1.inverse_transform_scores(self._scores1)
+        scores2 = self.stacker2.inverse_transform_scores(self._scores2)
+        return scores1, scores2
+
+    def homogeneous_patterns(self, correction=None, alpha=0.05):
+        '''Return the homogeneous patterns of the left and right field.
+
+        The homogeneous patterns are the correlation coefficients between the 
+        input data and the scores.
+
+        More precisely, the homogeneous patterns `r_{hom}` are defined as
+
+        .. math::
+          r_{hom, x} = \\corr \\left(X, A_x \\right)
+        .. math::
+          r_{hom, y} = \\corr \\left(Y, A_y \\right)
+
+        where :math:`X` and :math:`Y` are the input data, :math:`A_x` and :math:`A_y`
+        are the scores of the left and right field, respectively.
+
+        Parameters:
+        -------------
+        correction: str, default=None
+            Method to apply a multiple testing correction. If None, no correction
+            is applied.  Available methods are:
+            - bonferroni : one-step correction
+            - sidak : one-step correction
+            - holm-sidak : step down method using Sidak adjustments
+            - holm : step-down method using Bonferroni adjustments
+            - simes-hochberg : step-up method (independent)
+            - hommel : closed method based on Simes tests (non-negative)
+            - fdr_bh : Benjamini/Hochberg (non-negative) (default)
+            - fdr_by : Benjamini/Yekutieli (negative)
+            - fdr_tsbh : two stage fdr correction (non-negative)
+            - fdr_tsbky : two stage fdr correction (non-negative)
+        alpha: float, default=0.05
+            The desired family-wise error rate. Not used if `correction` is None.
+
+        Returns:
+        ----------
+        patterns1: DataArray | Dataset | List[DataArray]
+            Left homogenous patterns.
+        patterns2: DataArray | Dataset | List[DataArray]
+            Right homogenous patterns.
+        pvals1: DataArray | Dataset | List[DataArray]
+            Left p-values.
+        pvals2: DataArray | Dataset | List[DataArray]
+            Right p-values.
+
+        '''
+        patterns1, pvals1 = pearson_correlation(self.data1, self._scores1, correction=correction, alpha=alpha)
+        patterns2, pvals2 = pearson_correlation(self.data2, self._scores2, correction=correction, alpha=alpha)
+
+        patterns1 = self.stacker1.inverse_transform_data(self._singular_vectors1)
+        patterns2 = self.stacker2.inverse_transform_data(self._singular_vectors2)
+
+        return patterns1, patterns2, pvals1, pvals2
+
+    def heterogeneous_patterns(self, correction=None, alpha=0.05):
+        '''Return the heterogeneous patterns of the left and right field.
+        
+        The heterogeneous patterns are the correlation coefficients between the
+        input data and the scores of the other field.
+        
+        More precisely, the heterogeneous patterns `r_{het}` are defined as
+        
+        .. math::
+          r_{het, x} = \\corr \\left(X, A_y \\right)
+        .. math::
+          r_{het, y} = \\corr \\left(Y, A_x \\right)
+        
+        where :math:`X` and :math:`Y` are the input data, :math:`A_x` and :math:`A_y`
+        are the scores of the left and right field, respectively.
+
+        Parameters:
+        -------------
+        correction: str, default=None
+            Method to apply a multiple testing correction. If None, no correction
+            is applied.  Available methods are: 
+            - bonferroni : one-step correction
+            - sidak : one-step correction
+            - holm-sidak : step down method using Sidak adjustments
+            - holm : step-down method using Bonferroni adjustments
+            - simes-hochberg : step-up method (independent)
+            - hommel : closed method based on Simes tests (non-negative)
+            - fdr_bh : Benjamini/Hochberg (non-negative) (default)
+            - fdr_by : Benjamini/Yekutieli (negative)
+            - fdr_tsbh : two stage fdr correction (non-negative)
+            - fdr_tsbky : two stage fdr correction (non-negative)
+        alpha: float, default=0.05
+            The desired family-wise error rate. Not used if `correction` is None.
+
+        '''
+        patterns1, pvals1 = pearson_correlation(self.data1, self._scores2, correction=correction, alpha=alpha)
+        patterns2, pvals2 = pearson_correlation(self.data2, self._scores1, correction=correction, alpha=alpha)
+
+        patterns1 = self.stacker1.inverse_transform_data(self._singular_vectors2)
+        patterns2 = self.stacker2.inverse_transform_data(self._singular_vectors1)
+
+        return patterns1, patterns2, pvals1, pvals2
+
+    def compute(self):
+        '''Computing the model will compute and load all DaskArrays.'''
+
+        self._singular_values = self._singular_values.compute()
+        self._explained_variance = self._explained_variance.compute()
+        self._squared_total_variance = self._squared_total_variance.compute()
+        self._squared_covariance_fraction = self._squared_covariance_fraction.compute()
+        self._singular_vectors1 = self._singular_vectors1.compute()
+        self._singular_vectors2 = self._singular_vectors2.compute()
+        self._norm1 = self._norm1.compute()
+        self._norm2 = self._norm2.compute()
+        self._scores1 = self._scores1.compute()
+        self._scores2 = self._scores2.compute()
