--- xeofs/xarray/__init__.py
@@ -1,6 +1,0 @@
-from .eof import EOF
-from .rotator import Rotator
-from .bootstrapper import Bootstrapper
-from .mca import MCA
-from .mca_rotator import MCA_Rotator
-from .rock_pca import ROCK_PCA
